I changed the verification email template to have a button and i changed the api a little bit that shuld get a link (attached with token) instead of posting a token,
i also changed the car api to be based on company when adding, deleting, searchign, editing, i didn't want to change the model of database if i could help it

+++ backend/src/controllers/carController.js
@@ -4,9 +4,13 @@ import User from '../../models/user.js';
 export const addCar = async (req, res) => {
     console.log("req.body", req.body);
     
-
     try{
+        const userId = req.user.userId;
 
+        //find the company name of the user 
+        const user = await User.findById(userId);
+        if(!user) return res.status(404).json({message:'User not found'});
+        
         //Grab information from car:
         const {licensePlate, rentalStatus, currentRental, year, color, make, model, mileage, repairStatus, warningLightIndicators, vehicleIdentificationNumber, carType} = req.body;
 
@@ -16,16 +20,11 @@ export const addCar = async (req, res) => {
         ) {
             return res.status(400).json({message: "All fields are required"});
         }
-        //First get the proper user:
-        const userId = req.user.userId;
-
-        //find the company name of the user 
-        const user = await User.findById(userId);
-        if(!user) return res.status(404).json({message:'User not found'});
 
         //add the car data
         const car = new Car({
             userID: user._id,
+            companyName: user.companyName,
             licensePlate,
             year,
             color,
@@ -40,7 +39,7 @@ export const addCar = async (req, res) => {
 
         await car.save();
 
-        res.status(201).json ({ success: true, mesage: 'Car added successfully', car});
+        res.status(201).json ({ success: true, message: 'Car added successfully', car});
 
     }catch(error){
         console.error(error);
@@ -54,16 +53,16 @@ export const addCar = async (req, res) => {
 
 export const deleteCar = async(req, res) => {
     try{
-        const carId = req.params.id;
         const userId = req.user.userId;
+        const carId = req.params.id;
 
         if(!carId){
-            return res.status(400).json({success: false, messag: "Car ID is required"});
+            return res.status(400).json({success: false, message: "Car ID is required"});
         }
 
         //get the current user 
         const user = await User.findById(userId);
-        if(!user)return res.status(404).json ({sucess: false, message: "User was not found"});
+        if(!user)return res.status(404).json ({success: false, message: "User was not found"});
 
         //then find the car we want
         const car = await Car.findById(carId);
@@ -72,11 +71,10 @@ export const deleteCar = async(req, res) => {
         }
 
 
-        //find the company and compare, just in case, i commented out cause we should never be able to see cars we cant access
-        // const carOwner = await User.findById(car.userID);
-        // if(user.companyName !== carOwner.companyName){
-        //     return res.status(403).json({success: false, message: "You don't have permission to delete this car"});
-        // }
+        //find the company and compare
+        if(user.companyName !== car.companyName){
+            return res.status(403).json({success: false, message: "You don't have permission to delete this car"});
+        }
 
         await Car.findByIdAndDelete(carId);
         res.status(200).json ({success: true, message: "Car successfully deleted"});
@@ -99,9 +97,8 @@ export const editCar = async (req, res) => {
         const car = await Car.findById(carId);
         if(!car) return res.status(404).json({success:false, message: "Car not found"});
          
-        //extra protection to make sure outside of users of company can edit car
-        const carOwner = await User.findById(car.userID);
-        if(carOwner.companyName !== user.companyName) return res.status(403).json({success: false, message: "Access Denied "});
+        //extra protection to make sure outside of users of company can't edit car
+        if(car.companyName !== user.companyName) return res.status(403).json({success: false, message: "Access Denied "});
     
 
     const fieldsUpdate = [
@@ -125,7 +122,7 @@ export const editCar = async (req, res) => {
         });
 
         await car.save();
-        res.status(200).json({success: true, message: "Car updated successsfully", car});
+        res.status(200).json({success: true, message: "Car updated successfully", car});
 
     }catch (error){
         console.error(error);
@@ -143,7 +140,8 @@ export const getCars = async(req, res)=> {
         const user = await User.findById(userId);
         if(!user) return res.status(404).json({success: false, message: "User not found"});
 
-        let query = {userID: user._id};
+
+        let query = {companyName: user.companyName};
 
         //If we want to add filtering in our searches:
         if(req.query.rentalStatus) query.rentalStatus = req.query.rentalStatus;
